new services and repositories for item, history and user

+++ gameecommerce-backend/src/main/java/com/gameecommerce/backend/item/Item.java
@@ -20,7 +20,8 @@ public class Item {
     @Column(length = 4000)
     private String description;
 
-    private double price;
+    // in cents
+    private int price;
 
     private int stock;
